+++ Commands/Base/PipeBinds/AppMetadataPipeBind.cs
@@ -7,12 +7,13 @@ namespace SharePointPnP.PowerShell.Commands.Base.PipeBinds
 {
     public sealed class AppMetadataPipeBind
     {
-        private readonly AppMetadata _appMetadata;
         private readonly Guid _id;
+        private readonly string _title;
 
         public AppMetadataPipeBind(AppMetadata metadata)
         {
-            _appMetadata = metadata;
+            //_appMetadata = metadata;
+            _id = metadata.Id;
         }
 
 
@@ -23,20 +24,15 @@ namespace SharePointPnP.PowerShell.Commands.Base.PipeBinds
 
         public AppMetadataPipeBind(string id)
         {
-            _id = Guid.Parse(id);
+            if(!Guid.TryParse(id, out _id))
+            {
+                _title = id;
+            }
         }
 
-        public Guid GetId()
-        {
-            if (_appMetadata != null)
-            {
-                return _appMetadata.Id;
-            }
-            else
-            {
-                return _id;
-            }
-        }
+        public Guid Id => _id;
+
+        public string Title => _title;
     }
 
 }